--- Deployment/test.py
@@ -1,2 +1,0 @@
-from streamlit_drawable_canvas import st_canvas
-help(st_canvas)
