--- InpaintingModule/requirements.txt
@@ -1,15 +1,0 @@
-torch
-torchvision
-torch-fidelity
-torchsummary
-matplotlib
-numpy
-pandas
-pillow
-tqdm
-python-telegram-bot
-transformers
-timm
-onnx
-onnxruntime
-streamlit
